moved arch to MVVM
