+++ lib/features/dog_data/services/dog_data_service.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 import 'dart:io';
-import 'package:dog_app/3model/models/breeds.dart';
+import 'package:dog_app/features/dog_data/models/breeds.dart';
 import 'package:http/http.dart';
 
 class DogDataService {
@@ -23,8 +23,7 @@ class DogDataService {
     } catch (e) {
       print(e);
       
-      breeds = [];
-      return breeds;
+      return null;
     }
     return breeds;
   }
