+++ lib/features/favourite_breeds/viewmodel/favourite_breeds_viewmodel.dart
@@ -1,5 +1,5 @@
 import 'dart:convert';
-import 'package:dog_app/3model/models/breeds.dart';
+import 'package:dog_app/features/dog_data/models/breeds.dart';
 import 'package:flutter/foundation.dart';
 import 'package:shared_preferences/shared_preferences.dart';
