+++ lib/main.dart
@@ -1,8 +1,9 @@
-import '1view/pages/home page/home_page.dart';
-import '1view/themes.dart';
-import '2viewmodel/favourite_breeds_viewmodel.dart';
+import 'package:dog_app/features/dog_data/viewmodel/dog_data_viewmodel.dart';
 import 'package:provider/provider.dart';
 import 'package:flutter/material.dart';
+import 'features/favourite_breeds/viewmodel/favourite_breeds_viewmodel.dart';
+import 'view/pages/home page/home_page.dart';
+import 'view/themes.dart';
 
 void main() {
   runApp(MyApp());
@@ -14,6 +15,7 @@ class MyApp extends StatelessWidget {
     return MultiProvider(
       providers: [
         ChangeNotifierProvider(create: (context) => FavouriteBreeds()),
+        ChangeNotifierProvider(create: (context) => DogData()),
       ],
       child: MaterialApp(
         debugShowCheckedModeBanner: false,
