+++ lib/view/pages/favourite page/favourites_page.dart
@@ -1,7 +1,6 @@
-import 'package:dog_app/1view/widgets/main_header.dart';
-import 'package:dog_app/1view/widgets/main_list.dart';
-
-import '../../../2viewmodel/favourite_breeds_viewmodel.dart';
+import 'package:dog_app/view/widgets/main_header.dart';
+import 'package:dog_app/view/widgets/main_list.dart';
+import '../../../features/favourite_breeds/viewmodel/favourite_breeds_viewmodel.dart';
 import 'package:provider/provider.dart';
 import 'package:flutter/material.dart';
 
@@ -24,6 +23,7 @@ class FavouritesPage extends StatelessWidget {
                 breeds: Provider.of<FavouriteBreeds>(context, listen: true)
                     .favouriteBreeds,
                 removeButton: true,
+                hasData: true,
               )),
             ],
           ),
