+++ lib/view/pages/home detail page/home_detail_page.dart
@@ -1,5 +1,5 @@
-import 'package:dog_app/3model/models/breeds.dart';
-import '../../../2viewmodel/favourite_breeds_viewmodel.dart';
+import 'package:dog_app/features/dog_data/models/breeds.dart';
+import '../../../features/favourite_breeds/viewmodel/favourite_breeds_viewmodel.dart';
 import 'package:provider/provider.dart';
 import 'widgets/description.dart';
 import 'package:flutter/material.dart';
